flatten a bit down
